// TalentConnect-main/src/app/core/auth/models/user-credentials.model.ts
/**
 * User Credentials - Pour le login/register
 */

export interface UserCredentials {
  email: string;
  password: string;
}

export interface RegisterRequest {
  email: string;
  password: string;
  firstName: string;
  lastName: string;
}

export interface LoginResponse {
  user: {
    id: string;
    email: string;
    firstName: string;
    lastName: string;
    role: string;
  };
  accessToken: string;
  refreshToken?: string;
  expiresIn: number;
}